fix: return 409 when pushing duplicate version without overwrite param
Rubygems does not display errors when the response code is successful.
Previously you would only see the error message if you were doing a
verbose push.  It also gives the actor a false sense of success.

By changing the status code to 409 the error will be displayed to the
author.  This is not a functional change server-side.